+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/heal/HealOperation.java
@@ -1,0 +1,16 @@
+package org.betonquest.betonquest.quest.action.heal;
+
+/**
+ * The operation to perform on the player's health.
+ */
+public enum HealOperation {
+
+    /**
+     * Add something to the health of the player.
+     */
+    ADD,
+    /**
+     * Set the health of the player.
+     */
+    SET
+}