+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/heal/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.action Action} implementation to heal a player.
+ */
+package org.betonquest.betonquest.quest.action.heal;